Rewrite SQL queries when using PostgreSQL
All SQL DBs we used before (MySQL and SQLite) support ? syntax
for placeholders while Postgres does not.

So we just rewrite statements at runtime to use $N syntax used
by PostgreSQL.

This is incomplete hack (it will not handle a lot cases correctly)
but it is enough for us at the moment.